requirements.txt
requests
Pillow
datasets
tqdm
numpy
sentence-transformers
chromadb
transformers
rank_bm25
torch
matplotlib
